+++ YAM2E/Classes/Project.cs
@@ -9,4 +9,5 @@ public class Project
 
     public bool OptimizeObjectData { get; set; } = false;
     public bool useTilesets { get; set; } = true;
+    public string ProjectSpecificROM { get; set; } = "";
 }