+++ YAM2E/FORMS/ProjectSettings.cs
@@ -1,4 +1,5 @@
-﻿using System;
+﻿using LAMP.Classes;
+using System;
 using System.Collections.Generic;
 using System.ComponentModel;
 using System.Data;
@@ -17,6 +18,7 @@ public partial class ProjectSettings : Form
         InitializeComponent();
         chb_rmv_mt_o_list.Checked = Globals.LoadedProject.OptimizeObjectData;
         rbt_use_tilesets.Checked = Globals.LoadedProject.useTilesets;
+        txb_rom_path.Text = Globals.LoadedProject.ProjectSpecificROM;
     }
 
     private void chb_rmv_mt_o_list_CheckedChanged(object sender, EventArgs e)
@@ -40,5 +42,16 @@ public partial class ProjectSettings : Form
             Globals.LoadedProject.useTilesets = true;
             MainWindow.Current.SwitchTilesetOffsetMode();
         }
+    }
+
+    private void txb_rom_path_TextChanged(object sender, EventArgs e)
+    {
+        Globals.LoadedProject.ProjectSpecificROM = txb_rom_path.Text;
+    }
+
+    private void btn_select_rom_Click(object sender, EventArgs e)
+    {
+        string path = Editor.ShowOpenDialog("Metroid 2: Return of Samus ROM (*.gb)|*.gb");
+        if (path != String.Empty) txb_rom_path.Text = path;
     }
 }