+++ Changelog.md
@@ -1,6 +1,7 @@
 # Version B 5.0
 ### Released: XX.XX.2023
 **Changes:**
+ * Project specific base ROMs
 **Fixes:** 
  * TestROM: Missile values were capped at 0xFF
  * TestROM: Wrong math on the energy value